+++ core/src/main/java/com/google/zxing/common/DefaultGridSampler.java
@@ -55,7 +55,7 @@ public final class DefaultGridSampler extends GridSampler {
     float[] points = new float[2 * dimensionX];
     for (int y = 0; y < dimensionY; y++) {
       int max = points.length;
-      float iValue = (float) y + 0.5f;
+      float iValue = y + 0.5f;
       for (int x = 0; x < max; x += 2) {
         points[x] = (float) (x / 2) + 0.5f;
         points[x + 1] = iValue;
